+++ data-raw/calibrations-FLMS00440/save-calibs-FLMS00440.R
@@ -30,6 +30,7 @@ FLMS00440_simple.mthd <- list()
 # create an object with the parameters for simple method for Jaz
 # suitable for any light source, but not as good for irradiance.
 FLMS00440_none.mthd <- list(
+  name = "FLMS00440_none.mthd",
   spectrometer.sn = descriptor$spectrometer.sn,
   stray.light.method = "none",
   stray.light.wl = c(NA_real_, NA_real_),